[#3477] PR feedback
